feat: unify scaffold command with prompt
As discussed in a separate channel, this unifies the scaffold command into a
single prompt.  This prevents us from polluting the CLI commands with too many
different commands, and gives us the option to choose better names later.

It seems that this command will be renamed to init but that's out of scope
for this quick PR.